wip missing colors and images
