+++ libs/content/workspace/src/tab/svg_editor/renderer.rs
@@ -1,6 +1,8 @@
 use std::collections::HashMap;
 
 use glam::f64::DVec2;
+use lb_rs::svg::diff::DiffState;
+use lb_rs::svg::element::{Element, Image};
 use lb_rs::Uuid;
 use lyon::math::Point;
 use lyon::path::{AttributeIndex, LineCap, LineJoin};
@@ -13,7 +15,6 @@ use rayon::prelude::*;
 use resvg::usvg::{ImageKind, Transform};
 use tracing::{span, Level};
 
-use super::parser::{self};
 use super::Buffer;
 
 const STROKE_WIDTH: AttributeIndex = 0;
@@ -53,16 +54,17 @@ impl Renderer {
 
     pub fn render_svg(
         &mut self, ui: &mut egui::Ui, buffer: &mut Buffer, painter: &mut egui::Painter,
-    ) {
+    ) -> DiffState {
         let frame = ui.ctx().frame_nr();
         let span = span!(Level::TRACE, "rendering svg", frame);
         let _ = span.enter();
+        let mut diff_state = DiffState::default();
 
         let dark_mode_changed = ui.visuals().dark_mode != self.dark_mode;
         self.dark_mode = ui.visuals().dark_mode;
 
-        // todo: should avoid running this on every frame, because the images are allocated once
-        load_image_textures(buffer, ui);
+        // todo: i need to allocate the textures on another struct instead of using the image
+        // load_image_textures(buffer, ui);
 
         let paint_ops: Vec<(Uuid, RenderOp)> = buffer
             .elements
@@ -95,12 +97,15 @@ impl Renderer {
         for (id, paint_op) in paint_ops {
             match paint_op {
                 RenderOp::Delete => {
+                    diff_state.delete_changed = true;
                     self.mesh_cache.remove(&id);
                 }
                 RenderOp::Paint(m) => {
+                    diff_state.data_changed = true;
                     self.mesh_cache.insert(id.to_owned(), m);
                 }
                 RenderOp::Transform(t) => {
+                    diff_state.transformed = Some(t);
                     if let Some(egui::Shape::Mesh(m)) = self.mesh_cache.get_mut(&id) {
                         for v in &mut m.vertices {
                             v.pos.x = t.sx * v.pos.x + t.tx;
@@ -112,7 +117,12 @@ impl Renderer {
         }
 
         if !self.mesh_cache.is_empty() {
-            painter.extend(buffer.elements.iter().rev().filter_map(|(id, _)| {
+            painter.extend(buffer.elements.iter_mut().rev().filter_map(|(id, el)| {
+                match el {
+                    Element::Path(p) => p.diff_state = DiffState::default(),
+                    Element::Image(i) => i.diff_state = DiffState::default(),
+                    Element::Text(_) => todo!(),
+                }
                 if let Some(egui::Shape::Mesh(m)) = self.mesh_cache.get(id) {
                     if !m.vertices.is_empty() && !m.indices.is_empty() {
                         Some(egui::Shape::mesh(m.to_owned()))
@@ -123,46 +133,48 @@ impl Renderer {
                     None
                 }
             }));
-        }
+        };
+
+        diff_state
     }
 }
 
-fn load_image_textures(buffer: &mut Buffer, ui: &mut egui::Ui) {
-    for (id, el) in buffer.elements.iter_mut() {
-        if let parser::Element::Image(img) = el {
-            match &img.data {
-                ImageKind::JPEG(bytes) | ImageKind::PNG(bytes) => {
-                    let image = image::load_from_memory(bytes).unwrap();
+// fn load_image_textures(buffer: &mut Buffer, ui: &mut egui::Ui) {
+//     for (id, el) in buffer.elements.iter_mut() {
+//         if let Element::Image(img) = el {
+//             match &img.data {
+//                 ImageKind::JPEG(bytes) | ImageKind::PNG(bytes) => {
+//                     let image = image::load_from_memory(bytes).unwrap();
 
-                    let egui_image = egui::ColorImage::from_rgba_unmultiplied(
-                        [image.width() as usize, image.height() as usize],
-                        &image.to_rgba8(),
-                    );
+//                     let egui_image = egui::ColorImage::from_rgba_unmultiplied(
+//                         [image.width() as usize, image.height() as usize],
+//                         &image.to_rgba8(),
+//                     );
 
-                    if img.texture.is_none() {
-                        img.texture = Some(ui.ctx().load_texture(
-                            format!("canvas_img_{}", id),
-                            egui_image,
-                            egui::TextureOptions::LINEAR,
-                        ));
-                    }
-                }
-                ImageKind::GIF(_) => todo!(),
-                ImageKind::SVG(_) => todo!(),
-            }
-        }
-    }
-}
+//                     if img.texture.is_none() {
+//                         img.texture = Some(ui.ctx().load_texture(
+//                             format!("canvas_img_{}", id),
+//                             egui_image,
+//                             egui::TextureOptions::LINEAR,
+//                         ));
+//                     }
+//                 }
+//                 ImageKind::GIF(_) => todo!(),
+//                 ImageKind::SVG(_) => todo!(),
+//             }
+//         }
+//     }
+// }
 
 // todo: maybe impl this on element struct
 fn tesselate_element(
-    el: &mut parser::Element, id: &Uuid, dark_mode: bool, frame: u64, master_transform: Transform,
+    el: &mut Element, id: &Uuid, dark_mode: bool, frame: u64, master_transform: Transform,
 ) -> Option<(Uuid, RenderOp)> {
     let mut mesh: VertexBuffers<_, u32> = VertexBuffers::new();
     let mut stroke_tess = StrokeTessellator::new();
 
     match el {
-        parser::Element::Path(p) => {
+        Element::Path(p) => {
             let span = span!(Level::TRACE, "tessellating path", frame = frame);
             let _ = span.enter();
 
@@ -170,8 +182,15 @@ fn tesselate_element(
                 if p.data.is_empty() {
                     return Some((*id, RenderOp::Delete));
                 }
-                let stroke_color = if dark_mode { stroke.color.1 } else { stroke.color.0 }
-                    .gamma_multiply(p.opacity);
+
+                // todo: change strokes based on dynamic / light mode
+                let stroke_color = egui::Color32::from_rgb(
+                    stroke.color.red,
+                    stroke.color.green,
+                    stroke.color.blue,
+                )
+                .gamma_multiply(stroke.opacity)
+                .gamma_multiply(p.opacity);
 
                 let mut builder = lyon::path::BuilderWithAttributes::new(1);
 
@@ -226,8 +245,9 @@ fn tesselate_element(
                 None
             }
         }
-        parser::Element::Image(img) => render_image(img, id),
-        parser::Element::Text(_) => todo!(),
+        // todo: draw images
+        Element::Image(_) => todo!(),
+        Element::Text(_) => todo!(),
     }
 }
 
@@ -235,27 +255,27 @@ fn devc_to_point(dvec: DVec2) -> Point {
     Point::new(dvec.x as f32, dvec.y as f32)
 }
 
-fn render_image(img: &mut parser::Image, id: &Uuid) -> Option<(Uuid, RenderOp)> {
-    match &img.data {
-        ImageKind::JPEG(_) | ImageKind::PNG(_) => {
-            if let Some(texture) = &img.texture {
-                let rect = egui::Rect {
-                    min: egui::pos2(img.view_box.rect.left(), img.view_box.rect.top()),
-                    max: egui::pos2(img.view_box.rect.right(), img.view_box.rect.bottom()),
-                };
-                let uv = egui::Rect {
-                    min: egui::Pos2 { x: 0.0, y: 0.0 },
-                    max: egui::Pos2 { x: 1.0, y: 1.0 },
-                };
+// fn render_image(img: &mut Image, id: &Uuid) -> Option<(Uuid, RenderOp)> {
+//     match &img.data {
+//         ImageKind::JPEG(_) | ImageKind::PNG(_) => {
+//             if let Some(texture) = &img.texture {
+//                 let rect = egui::Rect {
+//                     min: egui::pos2(img.view_box.rect.left(), img.view_box.rect.top()),
+//                     max: egui::pos2(img.view_box.rect.right(), img.view_box.rect.bottom()),
+//                 };
+//                 let uv = egui::Rect {
+//                     min: egui::Pos2 { x: 0.0, y: 0.0 },
+//                     max: egui::Pos2 { x: 1.0, y: 1.0 },
+//                 };
 
-                let mut mesh = egui::Mesh::with_texture(texture.id());
-                mesh.add_rect_with_uv(rect, uv, egui::Color32::WHITE.linear_multiply(img.opacity));
-                Some((*id, RenderOp::Paint(egui::Shape::mesh(mesh))))
-            } else {
-                None
-            }
-        }
-        ImageKind::GIF(_) => todo!(),
-        ImageKind::SVG(_) => todo!(),
-    }
-}
+//                 let mut mesh = egui::Mesh::with_texture(texture.id());
+//                 mesh.add_rect_with_uv(rect, uv, egui::Color32::WHITE.linear_multiply(img.opacity));
+//                 Some((*id, RenderOp::Paint(egui::Shape::mesh(mesh))))
+//             } else {
+//                 None
+//             }
+//         }
+//         ImageKind::GIF(_) => todo!(),
+//         ImageKind::SVG(_) => todo!(),
+//     }
+// }
