+++ libs/content/workspace/src/tab/svg_editor/selection.rs
@@ -1,14 +1,17 @@
 use bezier_rs::Subpath;
 use egui_animation::{animate_bool_eased, easing};
 use glam::{DAffine2, DMat2, DVec2};
-use lb_rs::Uuid;
+use lb_rs::{
+    svg::element::{Element, ManipulatorGroupId},
+    Uuid,
+};
 use resvg::usvg::Transform;
 
 use crate::{theme::icons::Icon, widgets::Button};
 
 use super::{
+    element::BoundedElement,
     history::TransformElement,
-    parser::ManipulatorGroupId,
     toolbar::ToolContext,
     util::{is_multi_touch, pointer_intersects_element},
     Buffer, DeleteElement, Event,
@@ -348,13 +351,13 @@ impl Selection {
         laso_selected_elements
     }
 
-    fn el_intersects_laso(&mut self, el: &super::parser::Element) -> bool {
+    fn el_intersects_laso(&mut self, el: &Element) -> bool {
         let laso_rect = match self.laso_rect {
             Some(val) => val,
             None => return false,
         };
         match el {
-            super::parser::Element::Path(path) => {
+            Element::Path(path) => {
                 let path_rect = path.bounding_box();
                 if laso_rect.intersects(path_rect) {
                     let laso_bb = Subpath::new_rect(
@@ -371,11 +374,11 @@ impl Selection {
                     false
                 }
             }
-            super::parser::Element::Image(img) => {
+            Element::Image(img) => {
                 let img_bb = img.bounding_box();
                 laso_rect.contains_rect(img_bb) || laso_rect.intersects(img_bb)
             }
-            super::parser::Element::Text(_) => todo!(),
+            Element::Text(_) => todo!(),
         }
     }
