+++ libs/lb/lb-rs/Cargo.toml
@@ -21,6 +21,7 @@ native-tls = ["reqwest/native-tls"]
 [dependencies]
 base64 = "0.13.0"
 basic-human-duration = "0.2.0"
+bezier-rs = "0.2.0"
 bincode = "1.3.3"
 time = "0.3.20"
 diffy = "0.3.0"
@@ -28,6 +29,7 @@ reqwest = { version = "0.11.1", default-features = false, features = [
     "blocking",
     "json",
 ] }
+indexmap = { version = "2.5.0", features = ["rayon"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.44"
 uuid = { version = "1.2.2", features = ["v4", "serde"] }
@@ -53,6 +55,9 @@ flate2 = "1.0"
 bip39-dict = "0.1.3"
 similar = { version = "2.6.0", features = ["unicode"] }
 unicode-segmentation = "1.10.0"
+usvg = "0.41.0"
+glam = "0.22.0"
+
 
 [dev-dependencies]
 criterion = "0.5.1"
