+++ libs/lb/lb-rs/src/svg/mod.rs
@@ -1,0 +1,3 @@
+pub mod buffer;
+pub mod diff;
+pub mod element;
